call/cc transformations now produce a module
Instead of producing a file with a hard-coded interval, the call/cc-based
transformations now produce a module. We have a separate runner that
dynamically loads one of these modules and takes an optional yield interval.

Example:

    $ ./bin/compile --transform lazy program.js program.compiled.js
    $ ./bin/run --yield 30000 program.compiled.js

If the --yield argument is omitted, the program never yields.

Finally, ./bin/stopify no longer supports the call/cc based transformations.
We could have left it around, but I figure this is less cluttered.