Benchmarking cleanup (#108)
This commit should make benchmarking more reliable.

The main differences are these:

Subprocesses spawned to run individual benchmarks have a timeout. This seems to help.
The time between individual yields is now output if --variance is set.
Internally, the output csv files are built in a post-processing step instead of collected from the output of parallel. This is more robust. In addition, its easy to re-run failed benchmarks without messing up the gathered data.