lazy runtime: blow up when execution ends in restore mode
Execution should always end in normal mode (this.mode === true).
If we ever end in restore mode, then we failed to fully resume execution.
This is a common bug that is difficult to unit test, because it often
results in skipped assertions. So far, the only way to debug
these kinds of errors is to look at the console output.
If "BEGIN STOPIFY BENCHMARK RESULTS" doe≈s not print, then something
went wrong. This addresses that problem.