Scala compiler in its own cloud function
It turns out we can just upload our own copy of the JRE, etc. and use the CLI
to build ScalaJS. This is somewhat faster than using SBT and addresses the SBT
lockfile issue. An issue with approach is that we need hardcode the name of the
main method (i.e., Runner.main). We could write a little hacky script to detect
it if we want.