Major cleanup and some minor fixes
This is a monster commit that cleans up a lot of code and fixes one major bug:

When unit-testing in the browser, JavaScript errors were not being turned into test case failures. This is now fixed. This was masking a bug in the HOF test case. (We had the wrong flag.) This is also fixed.